react to codereview part two
