+++ src/JustEat.StatsD/V2/IStatsDTransportV2.cs
@@ -2,7 +2,7 @@ using System;
 
 namespace JustEat.StatsD.V2
 {
-    public interface IStatsDTransportV2
+    internal interface IStatsDTransportV2
     {
         void Send(ArraySegment<byte> metric);
     }
