+++ src/JustEat.StatsD/V2/StatsDMessageKind.cs
@@ -1,0 +1,9 @@
+namespace JustEat.StatsD.V2
+{
+    internal enum StatsDMessageKind
+    {
+        Counter,
+        Timing,
+        Gauge
+    }
+}
